Rename tcptx to tx and pdurx to rx
Looking at the lines of communication, tcprx ONLY
communicates with pdurx so there is no need for tcprx
to be visible outside that relationship.

Also both tcptx and pdurx use erlang records on their outside
facing interface, so it makes sense to rename both to simply
tx and rx to keep the interface symmetrical and simple
to understand.

That rx uses a TCPRX privately to communicate with the
network will become a private matter. This will happen
in the next few commits

Signed-off-by: Essien Ita Essien <essiene@gmail.com>